Extract Practice and Results screens out of App component
